+++ cmd/dep/status.go
@@ -9,6 +9,7 @@ import (
 	"encoding/json"
 	"flag"
 	"fmt"
+	"html/template"
 	"io"
 	"io/ioutil"
 	"log"
@@ -19,7 +20,6 @@ import (
 	"github.com/golang/dep"
 	"github.com/golang/dep/gps"
 	"github.com/golang/dep/gps/paths"
-	"github.com/golang/dep/gps/pkgtree"
 	"github.com/pkg/errors"
 )
 
@@ -85,26 +85,27 @@ type statusCommand struct {
 }
 
 type outputter interface {
-	BasicHeader()
-	BasicLine(*BasicStatus)
-	BasicFooter()
-	MissingHeader()
-	MissingLine(*MissingStatus)
-	MissingFooter()
+	BasicHeader() error
+	BasicLine(*BasicStatus) error
+	BasicFooter() error
+	MissingHeader() error
+	MissingLine(*MissingStatus) error
+	MissingFooter() error
 }
 
 type tableOutput struct{ w *tabwriter.Writer }
 
-func (out *tableOutput) BasicHeader() {
-	fmt.Fprintf(out.w, "PROJECT\tCONSTRAINT\tVERSION\tREVISION\tLATEST\tPKGS USED\n")
+func (out *tableOutput) BasicHeader() error {
+	_, err := fmt.Fprintf(out.w, "PROJECT\tCONSTRAINT\tVERSION\tREVISION\tLATEST\tPKGS USED\n")
+	return err
 }
 
-func (out *tableOutput) BasicFooter() {
-	out.w.Flush()
+func (out *tableOutput) BasicFooter() error {
+	return out.w.Flush()
 }
 
-func (out *tableOutput) BasicLine(bs *BasicStatus) {
-	fmt.Fprintf(out.w,
+func (out *tableOutput) BasicLine(bs *BasicStatus) error {
+	_, err := fmt.Fprintf(out.w,
 		"%s\t%s\t%s\t%s\t%s\t%d\t\n",
 		bs.ProjectRoot,
 		bs.getConsolidatedConstraint(),
@@ -113,22 +114,25 @@ func (out *tableOutput) BasicLine(bs *BasicStatus) {
 		bs.getConsolidatedLatest(shortRev),
 		bs.PackageCount,
 	)
+	return err
 }
 
-func (out *tableOutput) MissingHeader() {
-	fmt.Fprintln(out.w, "PROJECT\tMISSING PACKAGES")
+func (out *tableOutput) MissingHeader() error {
+	_, err := fmt.Fprintln(out.w, "PROJECT\tMISSING PACKAGES")
+	return err
 }
 
-func (out *tableOutput) MissingLine(ms *MissingStatus) {
-	fmt.Fprintf(out.w,
+func (out *tableOutput) MissingLine(ms *MissingStatus) error {
+	_, err := fmt.Fprintf(out.w,
 		"%s\t%s\t\n",
 		ms.ProjectRoot,
 		ms.MissingPackages,
 	)
+	return err
 }
 
-func (out *tableOutput) MissingFooter() {
-	out.w.Flush()
+func (out *tableOutput) MissingFooter() error {
+	return out.w.Flush()
 }
 
 type jsonOutput struct {
@@ -137,28 +141,32 @@ type jsonOutput struct {
 	missing []*MissingStatus
 }
 
-func (out *jsonOutput) BasicHeader() {
+func (out *jsonOutput) BasicHeader() error {
 	out.basic = []*rawStatus{}
+	return nil
 }
 
-func (out *jsonOutput) BasicFooter() {
-	json.NewEncoder(out.w).Encode(out.basic)
+func (out *jsonOutput) BasicFooter() error {
+	return json.NewEncoder(out.w).Encode(out.basic)
 }
 
-func (out *jsonOutput) BasicLine(bs *BasicStatus) {
+func (out *jsonOutput) BasicLine(bs *BasicStatus) error {
 	out.basic = append(out.basic, bs.marshalJSON())
+	return nil
 }
 
-func (out *jsonOutput) MissingHeader() {
+func (out *jsonOutput) MissingHeader() error {
 	out.missing = []*MissingStatus{}
+	return nil
 }
 
-func (out *jsonOutput) MissingLine(ms *MissingStatus) {
+func (out *jsonOutput) MissingLine(ms *MissingStatus) error {
 	out.missing = append(out.missing, ms)
+	return nil
 }
 
-func (out *jsonOutput) MissingFooter() {
-	json.NewEncoder(out.w).Encode(out.missing)
+func (out *jsonOutput) MissingFooter() error {
+	return json.NewEncoder(out.w).Encode(out.missing)
 }
 
 type dotOutput struct {
@@ -168,28 +176,51 @@ type dotOutput struct {
 	p *dep.Project
 }
 
-func (out *dotOutput) BasicHeader() {
+func (out *dotOutput) BasicHeader() error {
 	out.g = new(graphviz).New()
 
-	ptree, _ := out.p.ParseRootPackageTree()
+	ptree, err := out.p.ParseRootPackageTree()
 	// TODO(sdboyer) should be true, true, false, out.p.Manifest.IgnoredPackages()
 	prm, _ := ptree.ToReachMap(true, false, false, nil)
 
 	out.g.createNode(string(out.p.ImportRoot), "", prm.FlattenFn(paths.IsStandardImportPath))
+
+	return err
 }
 
-func (out *dotOutput) BasicFooter() {
+func (out *dotOutput) BasicFooter() error {
 	gvo := out.g.output()
-	fmt.Fprintf(out.w, gvo.String())
+	_, err := fmt.Fprintf(out.w, gvo.String())
+	return err
 }
 
-func (out *dotOutput) BasicLine(bs *BasicStatus) {
+func (out *dotOutput) BasicLine(bs *BasicStatus) error {
 	out.g.createNode(bs.ProjectRoot, bs.getConsolidatedVersion(), bs.Children)
+	return nil
 }
 
-func (out *dotOutput) MissingHeader()                {}
-func (out *dotOutput) MissingLine(ms *MissingStatus) {}
-func (out *dotOutput) MissingFooter()                {}
+func (out *dotOutput) MissingHeader() error                { return nil }
+func (out *dotOutput) MissingLine(ms *MissingStatus) error { return nil }
+func (out *dotOutput) MissingFooter() error                { return nil }
+
+type templateOutput struct {
+	w    io.Writer
+	tmpl *template.Template
+}
+
+func (out *templateOutput) BasicHeader() error { return nil }
+func (out *templateOutput) BasicFooter() error { return nil }
+
+func (out *templateOutput) BasicLine(bs *BasicStatus) error {
+	return out.tmpl.Execute(out.w, bs)
+}
+
+func (out *templateOutput) MissingHeader() error { return nil }
+func (out *templateOutput) MissingFooter() error { return nil }
+
+func (out *templateOutput) MissingLine(ms *MissingStatus) error {
+	return out.tmpl.Execute(out.w, ms)
+}
 
 func (cmd *statusCommand) Run(ctx *dep.Ctx, args []string) error {
 	p, err := ctx.LoadProject()
@@ -230,6 +261,15 @@ func (cmd *statusCommand) Run(ctx *dep.Ctx, args []string) error {
 			p: p,
 			o: cmd.output,
 			w: &buf,
+		}
+	case cmd.template != "":
+		tmpl, err := template.New("status").Parse(cmd.template)
+		if err != nil {
+			return err
+		}
+		out = &templateOutput{
+			w:    &buf,
+			tmpl: tmpl,
 		}
 	default:
 		out = &tableOutput{
@@ -389,7 +429,7 @@ func runStatusAll(ctx *dep.Ctx, out outputter, p *dep.Project, sm gps.SourceMana
 		return false, 0, errors.Wrapf(err, "could not set up solver for input hashing")
 	}
 
-	cm := collectConstraints(ptree, p, sm)
+	cm := collectConstraints(ctx, p, sm)
 
 	// Get the project list and sort it so that the printed output users see is
 	// deterministically ordered. (This may be superfluous if the lock is always
@@ -404,7 +444,9 @@ func runStatusAll(ctx *dep.Ctx, out outputter, p *dep.Project, sm gps.SourceMana
 		// complete picture of all deps. That eliminates the need for at least
 		// some checks.
 
-		out.BasicHeader()
+		if err := out.BasicHeader(); err != nil {
+			return false, 0, err
+		}
 
 		logger.Println("Checking upstream projects:")
 
@@ -461,7 +503,7 @@ func runStatusAll(ctx *dep.Ctx, out outputter, p *dep.Project, sm gps.SourceMana
 				} else {
 					bs.Constraint = gps.Any()
 					for _, c := range cm[bs.ProjectRoot] {
-						bs.Constraint = c.Intersect(bs.Constraint)
+						bs.Constraint = c.Constraint.Intersect(bs.Constraint)
 					}
 				}
 
@@ -470,7 +512,12 @@ func runStatusAll(ctx *dep.Ctx, out outputter, p *dep.Project, sm gps.SourceMana
 				if bs.Version != nil && bs.Version.Type() != gps.IsVersion {
 					c, has := p.Manifest.Constraints[proj.Ident().ProjectRoot]
 					if !has {
-						c.Constraint = gps.Any()
+						// Get constraint for locked project
+						for _, lockedP := range p.Lock.P {
+							if lockedP.Ident().ProjectRoot == proj.Ident().ProjectRoot {
+								c.Constraint = lockedP.Version()
+							}
+						}
 					}
 					// TODO: This constraint is only the constraint imposed by the
 					// current project, not by any transitive deps. As a result,
@@ -552,10 +599,14 @@ func runStatusAll(ctx *dep.Ctx, out outputter, p *dep.Project, sm gps.SourceMana
 
 		// Use the collected BasicStatus in outputter.
 		for _, proj := range slp {
-			out.BasicLine(bsMap[string(proj.Ident().ProjectRoot)])
+			if err := out.BasicLine(bsMap[string(proj.Ident().ProjectRoot)]); err != nil {
+				return false, 0, err
+			}
 		}
 
-		out.BasicFooter()
+		if footerErr := out.BasicFooter(); footerErr != nil {
+			return false, 0, footerErr
+		}
 
 		return false, errCount, err
 	}
@@ -601,7 +652,9 @@ func runStatusAll(ctx *dep.Ctx, out outputter, p *dep.Project, sm gps.SourceMana
 		return false, 0, errors.New("address issues with undeducible import paths to get more status information")
 	}
 
-	out.MissingHeader()
+	if err = out.MissingHeader(); err != nil {
+		return false, 0, err
+	}
 
 outer:
 	for root, pkgs := range roots {
@@ -614,9 +667,14 @@ outer:
 		}
 
 		hasMissingPkgs = true
-		out.MissingLine(&MissingStatus{ProjectRoot: string(root), MissingPackages: pkgs})
+		err := out.MissingLine(&MissingStatus{ProjectRoot: string(root), MissingPackages: pkgs})
+		if err != nil {
+			return false, 0, err
+		}
 	}
-	out.MissingFooter()
+	if err = out.MissingFooter(); err != nil {
+		return false, 0, err
+	}
 
 	// We are here because of an input-digest mismatch. Return error.
 	return hasMissingPkgs, 0, errInputDigestMismatch
@@ -639,7 +697,49 @@ func formatVersion(v gps.Version) string {
 	return v.String()
 }
 
-func collectConstraints(ptree pkgtree.PackageTree, p *dep.Project, sm gps.SourceManager) map[string][]gps.Constraint {
-	// TODO
-	return map[string][]gps.Constraint{}
+// projectConstraint stores ProjectRoot and Constraint for that project.
+type projectConstraint struct {
+	Project    gps.ProjectRoot
+	Constraint gps.Constraint
+}
+
+// constraintsCollection is a map of ProjectRoot(dependency) and a collection of
+// projectConstraint for the dependencies. This can be used to find constraints
+// on a dependency and the projects that apply those constraints.
+type constraintsCollection map[string][]projectConstraint
+
+// collectConstraints collects constraints declared by all the dependencies.
+func collectConstraints(ctx *dep.Ctx, p *dep.Project, sm gps.SourceManager) constraintsCollection {
+	constraintCollection := make(constraintsCollection)
+
+	// Get direct deps of the root project.
+	_, directDeps, err := getDirectDependencies(sm, p)
+	if err != nil {
+		ctx.Err.Println("Error getting direct deps:", err)
+	}
+	// Create a root analyzer.
+	rootAnalyzer := newRootAnalyzer(true, ctx, directDeps, sm)
+
+	// Iterate through the locked projects and collect constraints of all the projects.
+	for _, proj := range p.Lock.Projects() {
+		manifest, _, err := sm.GetManifestAndLock(proj.Ident(), proj.Version(), rootAnalyzer)
+		if err != nil {
+			ctx.Err.Println("Error getting manifest and lock:", err)
+			continue
+		}
+
+		// Get project constraints.
+		pc := manifest.DependencyConstraints()
+
+		// Iterate through the project constraints to get individual dependency
+		// project and constraint values.
+		for pr, pp := range pc {
+			constraintCollection[string(pr)] = append(
+				constraintCollection[string(pr)],
+				projectConstraint{proj.Ident().ProjectRoot, pp.Constraint},
+			)
+		}
+	}
+
+	return constraintCollection
 }