+++ manifest.go
@@ -24,26 +24,43 @@ const ManifestName = "Gopkg.toml"
 
 // Errors
 var (
-	errInvalidConstraint  = errors.New("\"constraint\" must be a TOML array of tables")
-	errInvalidOverride    = errors.New("\"override\" must be a TOML array of tables")
-	errInvalidRequired    = errors.New("\"required\" must be a TOML list of strings")
-	errInvalidIgnored     = errors.New("\"ignored\" must be a TOML list of strings")
+	errInvalidConstraint   = errors.Errorf("%q must be a TOML array of tables", "constraint")
+	errInvalidOverride     = errors.Errorf("%q must be a TOML array of tables", "override")
+	errInvalidRequired     = errors.Errorf("%q must be a TOML list of strings", "required")
+	errInvalidIgnored      = errors.Errorf("%q must be a TOML list of strings", "ignored")
+	errInvalidPrune        = errors.Errorf("%q must be a TOML table of booleans", "prune")
+	errInvalidPruneProject = errors.Errorf("%q must be a TOML array of tables", "prune.project")
+	errInvalidMetadata     = errors.New("metadata should be a TOML table")
+
 	errInvalidProjectRoot = errors.New("ProjectRoot name validation failed")
+
+	errInvalidPruneValue = errors.New("prune options values must be booleans")
+	errPruneSubProject   = errors.New("prune projects should not contain sub projects")
+
+	errRootPruneContainsName   = errors.Errorf("%q should not include a name", "prune")
+	errInvalidRootPruneValue   = errors.New("root prune options must be omitted instead of being set to false")
+	errInvalidPruneProjectName = errors.Errorf("%q in %q must be a string", "name", "prune.project")
+	errNoName                  = errors.New("no name provided")
 )
 
 // Manifest holds manifest file data and implements gps.RootManifest.
 type Manifest struct {
 	Constraints gps.ProjectConstraints
 	Ovr         gps.ProjectConstraints
-	Ignored     []string
-	Required    []string
+
+	Ignored  []string
+	Required []string
+
+	PruneOptions        gps.PruneOptions
+	PruneProjectOptions gps.PruneProjectOptions
 }
 
 type rawManifest struct {
-	Constraints []rawProject `toml:"constraint,omitempty"`
-	Overrides   []rawProject `toml:"override,omitempty"`
-	Ignored     []string     `toml:"ignored,omitempty"`
-	Required    []string     `toml:"required,omitempty"`
+	Constraints  []rawProject    `toml:"constraint,omitempty"`
+	Overrides    []rawProject    `toml:"override,omitempty"`
+	Ignored      []string        `toml:"ignored,omitempty"`
+	Required     []string        `toml:"required,omitempty"`
+	PruneOptions rawPruneOptions `toml:"prune,omitempty"`
 }
 
 type rawProject struct {
@@ -54,11 +71,33 @@ type rawProject struct {
 	Source   string `toml:"source,omitempty"`
 }
 
-// NewManifest instantiates a new manifest.
+type rawPruneOptions struct {
+	UnusedPackages bool `toml:"unused-packages,omitempty"`
+	NonGoFiles     bool `toml:"non-go,omitempty"`
+	GoTests        bool `toml:"go-tests,omitempty"`
+
+	Projects []rawPruneProjectOptions `toml:"project,omitempty"`
+}
+
+type rawPruneProjectOptions struct {
+	Name           string `toml:"name"`
+	UnusedPackages bool   `toml:"unused-packages,omitempty"`
+	NonGoFiles     bool   `toml:"non-go,omitempty"`
+	GoTests        bool   `toml:"go-tests,omitempty"`
+}
+
+const (
+	pruneOptionUnusedPackages = "unused-packages"
+	pruneOptionGoTests        = "go-tests"
+	pruneOptionNonGo          = "non-go"
+)
+
+// NewManifest instantites a new manifest.
 func NewManifest() *Manifest {
 	return &Manifest{
-		Constraints: make(gps.ProjectConstraints),
-		Ovr:         make(gps.ProjectConstraints),
+		Constraints:  make(gps.ProjectConstraints),
+		Ovr:          make(gps.ProjectConstraints),
+		PruneOptions: gps.PruneNestedVendorDirs,
 	}
 }
 
@@ -67,7 +106,7 @@ func validateManifest(s string) ([]error, error) {
 	// Load the TomlTree from string
 	tree, err := toml.Load(s)
 	if err != nil {
-		return warns, errors.Wrap(err, "Unable to load TomlTree from string")
+		return warns, errors.Wrap(err, "unable to load TomlTree from string")
 	}
 	// Convert tree to a map
 	manifest := tree.ToMap()
@@ -80,7 +119,7 @@ func validateManifest(s string) ([]error, error) {
 		case "metadata":
 			// Check if metadata is of Map type
 			if reflect.TypeOf(val).Kind() != reflect.Map {
-				warns = append(warns, errors.New("metadata should be a TOML table"))
+				warns = append(warns, errInvalidMetadata)
 			}
 		case "constraint", "override":
 			valid := true
@@ -95,13 +134,17 @@ func validateManifest(s string) ([]error, error) {
 				if valid {
 					// Iterate through each array of tables
 					for _, v := range rawProj {
+						ruleProvided := false
+						props := v.(map[string]interface{})
 						// Check the individual field's key to be valid
-						for key, value := range v.(map[string]interface{}) {
+						for key, value := range props {
 							// Check if the key is valid
 							switch key {
-							case "name", "branch", "version", "source":
-								// valid key
+							case "name":
+							case "branch", "version", "source":
+								ruleProvided = true
 							case "revision":
+								ruleProvided = true
 								if valueStr, ok := value.(string); ok {
 									if abbrevRevHash.MatchString(valueStr) {
 										warns = append(warns, fmt.Errorf("revision %q should not be in abbreviated form", valueStr))
@@ -114,8 +157,13 @@ func validateManifest(s string) ([]error, error) {
 								}
 							default:
 								// unknown/invalid key
-								warns = append(warns, fmt.Errorf("Invalid key %q in %q", key, prop))
+								warns = append(warns, fmt.Errorf("invalid key %q in %q", key, prop))
 							}
+						}
+						if _, ok := props["name"]; !ok {
+							warns = append(warns, errNoName)
+						} else if !ruleProvided && prop == "constraint" {
+							warns = append(warns, fmt.Errorf("branch, version, revision, or source should be provided for %q", props["name"]))
 						}
 					}
 				}
@@ -151,12 +199,83 @@ func validateManifest(s string) ([]error, error) {
 					return warns, errInvalidRequired
 				}
 			}
+		case "prune":
+			pruneWarns, err := validatePruneOptions(val, true)
+			warns = append(warns, pruneWarns...)
+			if err != nil {
+				return warns, err
+			}
 		default:
-			warns = append(warns, fmt.Errorf("Unknown field in manifest: %v", prop))
+			warns = append(warns, fmt.Errorf("unknown field in manifest: %v", prop))
 		}
 	}
 
 	return warns, nil
+}
+
+func validatePruneOptions(val interface{}, root bool) (warns []error, err error) {
+	if reflect.TypeOf(val).Kind() != reflect.Map {
+		return warns, errInvalidPrune
+	}
+
+	for key, value := range val.(map[string]interface{}) {
+		switch key {
+		case pruneOptionNonGo, pruneOptionGoTests, pruneOptionUnusedPackages:
+			if option, ok := value.(bool); !ok {
+				return warns, errInvalidPruneValue
+			} else if root && !option {
+				return warns, errInvalidRootPruneValue
+			}
+		case "name":
+			if root {
+				warns = append(warns, errRootPruneContainsName)
+			} else if _, ok := value.(string); !ok {
+				return warns, errInvalidPruneProjectName
+			}
+		case "project":
+			if !root {
+				return warns, errPruneSubProject
+			}
+			if reflect.TypeOf(value).Kind() != reflect.Slice {
+				return warns, errInvalidPruneProject
+			}
+
+			for _, project := range value.([]interface{}) {
+				projectWarns, err := validatePruneOptions(project, false)
+				warns = append(warns, projectWarns...)
+				if err != nil {
+					return nil, err
+				}
+			}
+
+		default:
+			if root {
+				warns = append(warns, errors.Errorf("unknown field %q in %q", key, "prune"))
+			} else {
+				warns = append(warns, errors.Errorf("unknown field %q in %q", key, "prune.project"))
+			}
+		}
+	}
+
+	return warns, err
+}
+
+func checkRedundantPruneOptions(raw rawManifest) (warns []error) {
+	rootOptions := raw.PruneOptions
+
+	for _, project := range raw.PruneOptions.Projects {
+		if rootOptions.GoTests && project.GoTests {
+			warns = append(warns, errors.Errorf("redundant prune option %q set for %q", pruneOptionGoTests, project.Name))
+		}
+		if rootOptions.NonGoFiles && project.NonGoFiles {
+			warns = append(warns, errors.Errorf("redundant prune option %q set for %q", pruneOptionNonGo, project.Name))
+		}
+		if rootOptions.UnusedPackages && project.UnusedPackages {
+			warns = append(warns, errors.Errorf("redundant prune option %q set for %q", pruneOptionUnusedPackages, project.Name))
+		}
+	}
+
+	return warns
 }
 
 // ValidateProjectRoots validates the project roots present in manifest.
@@ -184,6 +303,10 @@ func ValidateProjectRoots(c *Ctx, m *Manifest, sm gps.SourceManager) error {
 		wg.Add(1)
 		go validate(pr)
 	}
+	for pr := range m.PruneProjectOptions {
+		wg.Add(1)
+		go validate(pr)
+	}
 
 	wg.Wait()
 	close(errorCh)
@@ -206,19 +329,21 @@ func readManifest(r io.Reader) (*Manifest, []error, error) {
 	buf := &bytes.Buffer{}
 	_, err := buf.ReadFrom(r)
 	if err != nil {
-		return nil, nil, errors.Wrap(err, "Unable to read byte stream")
+		return nil, nil, errors.Wrap(err, "unable to read byte stream")
 	}
 
 	warns, err := validateManifest(buf.String())
 	if err != nil {
-		return nil, warns, errors.Wrap(err, "Manifest validation failed")
+		return nil, warns, errors.Wrap(err, "manifest validation failed")
 	}
 
 	raw := rawManifest{}
 	err = toml.Unmarshal(buf.Bytes(), &raw)
 	if err != nil {
-		return nil, warns, errors.Wrap(err, "Unable to parse the manifest as TOML")
+		return nil, warns, errors.Wrap(err, "unable to parse the manifest as TOML")
 	}
+
+	warns = append(warns, checkRedundantPruneOptions(raw)...)
 
 	m, err := fromRawManifest(raw)
 	return m, warns, err
@@ -254,7 +379,41 @@ func fromRawManifest(raw rawManifest) (*Manifest, error) {
 		m.Ovr[name] = prj
 	}
 
+	m.PruneOptions, m.PruneProjectOptions = fromRawPruneOptions(raw.PruneOptions)
+
 	return m, nil
+}
+
+func fromRawPruneOptions(raw rawPruneOptions) (gps.PruneOptions, gps.PruneProjectOptions) {
+	rootOptions := gps.PruneNestedVendorDirs
+	pruneProjects := make(gps.PruneProjectOptions)
+
+	if raw.UnusedPackages {
+		rootOptions |= gps.PruneUnusedPackages
+	}
+	if raw.GoTests {
+		rootOptions |= gps.PruneGoTestFiles
+	}
+	if raw.NonGoFiles {
+		rootOptions |= gps.PruneNonGoFiles
+	}
+
+	for _, p := range raw.Projects {
+		pr := gps.ProjectRoot(p.Name)
+		pruneProjects[pr] = gps.PruneNestedVendorDirs
+
+		if raw.UnusedPackages {
+			pruneProjects[pr] |= gps.PruneUnusedPackages
+		}
+		if raw.GoTests {
+			pruneProjects[pr] |= gps.PruneGoTestFiles
+		}
+		if raw.NonGoFiles {
+			pruneProjects[pr] |= gps.PruneNonGoFiles
+		}
+	}
+
+	return rootOptions, pruneProjects
 }
 
 // toProject interprets the string representations of project information held in
@@ -288,17 +447,27 @@ func toProject(raw rawProject) (n gps.ProjectRoot, pp gps.ProjectProperties, err
 	}
 
 	pp.Source = raw.Source
+
 	return n, pp, nil
+}
+
+// MarshalTOML serializes this manifest into TOML via an intermediate raw form.
+func (m *Manifest) MarshalTOML() ([]byte, error) {
+	raw := m.toRaw()
+	result, err := toml.Marshal(raw)
+	return result, errors.Wrap(err, "unable to marshal the lock to a TOML string")
 }
 
 // toRaw converts the manifest into a representation suitable to write to the manifest file
 func (m *Manifest) toRaw() rawManifest {
 	raw := rawManifest{
-		Constraints: make([]rawProject, 0, len(m.Constraints)),
-		Overrides:   make([]rawProject, 0, len(m.Ovr)),
-		Ignored:     m.Ignored,
-		Required:    m.Required,
+		Constraints:  make([]rawProject, 0, len(m.Constraints)),
+		Overrides:    make([]rawProject, 0, len(m.Ovr)),
+		Ignored:      m.Ignored,
+		Required:     m.Required,
+		PruneOptions: rawPruneOptions{},
 	}
+
 	for n, prj := range m.Constraints {
 		raw.Constraints = append(raw.Constraints, toRawProject(n, prj))
 	}
@@ -308,6 +477,8 @@ func (m *Manifest) toRaw() rawManifest {
 		raw.Overrides = append(raw.Overrides, toRawProject(n, prj))
 	}
 	sort.Sort(sortedRawProjects(raw.Overrides))
+
+	// TODO(ibrasho): write out prune options.
 
 	return raw
 }
@@ -327,13 +498,6 @@ func (s sortedRawProjects) Less(i, j int) bool {
 	}
 
 	return l.Source < r.Source
-}
-
-// MarshalTOML serializes this manifest into TOML via an intermediate raw form.
-func (m *Manifest) MarshalTOML() ([]byte, error) {
-	raw := m.toRaw()
-	result, err := toml.Marshal(raw)
-	return result, errors.Wrap(err, "Unable to marshal the lock to a TOML string")
 }
 
 func toRawProject(name gps.ProjectRoot, project gps.ProjectProperties) rawProject {
@@ -363,6 +527,7 @@ func toRawProject(name gps.ProjectRoot, project gps.ProjectProperties) rawProjec
 		// Has to be a semver range.
 		raw.Version = project.Constraint.ImpliedCaretString()
 	}
+
 	return raw
 }
 
@@ -406,4 +571,16 @@ func (m *Manifest) RequiredPackages() map[string]bool {
 	}
 
 	return mp
+}
+
+// PruneOptionsFor returns the prune options for the passed project root.
+//
+// It will return the root prune options if the project does not have specific
+// options or if it does not exists in the manifest.
+func (m *Manifest) PruneOptionsFor(pr gps.ProjectRoot) gps.PruneOptions {
+	if po, ok := m.PruneProjectOptions[pr]; ok {
+		return po
+	}
+
+	return m.PruneOptions
 }