+++ CHANGELOG.md
@@ -1,15 +1,28 @@
 # v0.3.3 (Unreleased)
 
+NEW FEATURES:
+* Add support for importing from [glock](https://github.com/robfig/glock) based projects (#1422).
+* Add support for importing from [govendor](https://github.com/kardianos/govendor)
+  based projects (#815).
+* Allow override of cache directory location using environment variable
+`DEPCACHEDIR`. ([#1234](https://github.com/golang/dep/pull/1234))
+
 BUG FIXES:
 
 * Releases targeting Windows now have a `.exe` suffix (#1291).
 * Adaptively recover from dirty and corrupted git repositories in cache (#1279).
 * Suppress git password prompts in more places (#1357).
+* Fix `-no-vendor` flag for `ensure -update` (#1361).
+* Validate `git ls-remote` output and ignore all malformed lines (#1379)
+* Support [gopkg.in version zero](http://labix.org/gopkg.in#VersionZero) (#1243).
 
 IMPROVEMENTS:
 
 * Log as dependencies are pre-fetched during dep init ([#1176](https://github.com/golang/dep/pull/1176)).
 * Make the gps package importable ([#1349](https://github.com/golang/dep/pull/1349)).
+* Improve file copy performance by not forcing a file sync (PR #1408).
+* Skip empty constraints during import ([#1414](https://github.com/golang/dep/pull/1349))
+* Handle errors when writing status output ([#1420](https://github.com/golang/dep/pull/1420))
 
 # v0.3.2